Moved AbstractDAOTest to dpc-common.  Made it more generic.  Cleaned up pom files.
